Make match players clickable
Use long data type for accountId and lower match result layouts' height

+++ app/src/main/java/com/example/dotametrics/data/model/players/peers/PeersResult.kt
@@ -3,7 +3,7 @@ package com.example.dotametrics.data.model.players.peers
 import com.google.gson.annotations.SerializedName
 
 data class PeersResult(
-    @SerializedName("account_id") var accountId: Int? = null,
+    @SerializedName("account_id") var accountId: Long? = null,
     @SerializedName("last_played") var lastPlayed: Long? = null,
     @SerializedName("with_win") var withWin: Int? = null,
     @SerializedName("with_games") var withGames: Int? = null,
