+++ app/src/main/java/com/example/dotametrics/data/model/search/SearchResult.kt
@@ -6,7 +6,7 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class SearchResult(
-    @SerializedName("account_id") var accountId: Int? = null,
+    @SerializedName("account_id") var accountId: Long? = null,
     @SerializedName("personaname") var personaname: String? = null,
     @SerializedName("avatarfull") var avatarfull: String? = null,
     @SerializedName("last_match_time") var lastMatchTime: String? = null,
